# fixtures/task-solutions/optimistic-locking-direct/internal/httpapi/handler.go
package httpapi

import (
	"context"
	"encoding/json"
	"errors"
	"io"
	"mime"
	"net/http"
	"strconv"
	"strings"

	"github.com/go-chi/chi/v5"

	"specs-dont-hallucinate/taskservice/internal/task"
)

const maxRequestBody = 1 << 20

type Service interface {
	Create(context.Context, string) (task.Task, error)
	Get(context.Context, string) (task.Task, error)
	List(context.Context) ([]task.Task, error)
	Delete(context.Context, string) error
	Update(context.Context, string, string, int64) (task.Task, error)
}

type Handler struct {
	service Service
}

type createTaskRequest struct {
	Title string `json:"title"`
}

type updateTaskRequest struct {
	Title string `json:"title"`
}

type taskResponse struct {
	ID        string `json:"id"`
	Title     string `json:"title"`
	CreatedAt string `json:"createdAt"`
	Version   int64  `json:"version"`
}

type taskListResponse struct {
	Items []taskResponse `json:"items"`
}

type problem struct {
	Type   string `json:"type"`
	Title  string `json:"title"`
	Status int    `json:"status"`
	Detail string `json:"detail"`
}

func RegisterRoutes(router chi.Router, service Service) {
	handler := &Handler{service: service}
	router.Post("/tasks", handler.createTask)
	router.Get("/tasks", handler.listTasks)
	router.Get("/tasks/{taskId}", handler.getTask)
	router.Put("/tasks/{taskId}", handler.updateTask)
	router.Delete("/tasks/{taskId}", handler.deleteTask)
}

func (h *Handler) createTask(w http.ResponseWriter, r *http.Request) {
	mediaType, _, err := mime.ParseMediaType(r.Header.Get("Content-Type"))
	if err != nil || mediaType != "application/json" {
		writeProblem(w, validationProblem())
		return
	}
	var request createTaskRequest
	if err := decodeJSON(w, r, &request); err != nil {
		writeProblem(w, validationProblem())
		return
	}
	result, err := h.service.Create(r.Context(), request.Title)
	if err != nil {
		writeServiceError(w, err)
		return
	}
	writeTask(w, http.StatusCreated, result)
}

func (h *Handler) getTask(w http.ResponseWriter, r *http.Request) {
	result, err := h.service.Get(r.Context(), chi.URLParam(r, "taskId"))
	if err != nil {
		writeServiceError(w, err)
		return
	}
	writeTask(w, http.StatusOK, result)
}

func (h *Handler) updateTask(w http.ResponseWriter, r *http.Request) {
	expectedVersion, status, ok := parseIfMatch(r.Header.Values("If-Match"))
	if !ok {
		if status == http.StatusPreconditionRequired {
			writeProblem(w, preconditionRequiredProblem())
		} else {
			writeProblem(w, validationProblem())
		}
		return
	}
	mediaType, _, err := mime.ParseMediaType(r.Header.Get("Content-Type"))
	if err != nil || mediaType != "application/json" {
		writeProblem(w, validationProblem())
		return
	}
	var request updateTaskRequest
	if err := decodeJSON(w, r, &request); err != nil {
		writeProblem(w, validationProblem())
		return
	}
	result, err := h.service.Update(r.Context(), chi.URLParam(r, "taskId"), request.Title, expectedVersion)
	if err != nil {
		writeServiceError(w, err)
		return
	}
	writeTask(w, http.StatusOK, result)
}

func parseIfMatch(values []string) (int64, int, bool) {
	if len(values) == 0 {
		return 0, http.StatusPreconditionRequired, false
	}
	if len(values) != 1 {
		return 0, http.StatusBadRequest, false
	}
	value := strings.Trim(values[0], " \t")
	if len(value) < 3 || value[0] != '"' || value[len(value)-1] != '"' {
		return 0, http.StatusBadRequest, false
	}
	digits := value[1 : len(value)-1]
	if digits[0] < '1' || digits[0] > '9' {
		return 0, http.StatusBadRequest, false
	}
	for index := 1; index < len(digits); index++ {
		if digits[index] < '0' || digits[index] > '9' {
			return 0, http.StatusBadRequest, false
		}
	}
	version, err := strconv.ParseInt(digits, 10, 64)
	if err != nil {
		return 0, http.StatusBadRequest, false
	}
	return version, 0, true
}

func (h *Handler) listTasks(w http.ResponseWriter, r *http.Request) {
	items, err := h.service.List(r.Context())
	if err != nil {
		writeServiceError(w, err)
		return
	}
	response := taskListResponse{Items: make([]taskResponse, len(items))}
	for index, item := range items {
		response.Items[index] = responseFromTask(item)
	}
	writeJSON(w, http.StatusOK, response)
}

func (h *Handler) deleteTask(w http.ResponseWriter, r *http.Request) {
	if err := h.service.Delete(r.Context(), chi.URLParam(r, "taskId")); err != nil {
		writeServiceError(w, err)
		return
	}
	w.WriteHeader(http.StatusNoContent)
}

func decodeJSON(w http.ResponseWriter, r *http.Request, destination any) error {
	r.Body = http.MaxBytesReader(w, r.Body, maxRequestBody)
	decoder := json.NewDecoder(r.Body)
	decoder.DisallowUnknownFields()
	if err := decoder.Decode(destination); err != nil {
		return err
	}
	if err := decoder.Decode(&struct{}{}); !errors.Is(err, io.EOF) {
		return errors.New("request body must contain one JSON value")
	}
	return nil
}

func responseFromTask(value task.Task) taskResponse {
	return taskResponse{
		ID:        value.ID,
		Title:     value.Title,
		CreatedAt: value.CreatedAt.UTC().Format("2006-01-02T15:04:05.000000Z"),
		Version:   value.Version,
	}
}

func writeServiceError(w http.ResponseWriter, err error) {
	switch {
	case errors.Is(err, task.ErrInvalid):
		writeProblem(w, validationProblem())
	case errors.Is(err, task.ErrNotFound):
		writeProblem(w, notFoundProblem())
	case errors.Is(err, task.ErrPreconditionFailed):
		writeProblem(w, preconditionFailedProblem())
	default:
		writeProblem(w, internalProblem())
	}
}

func validationProblem() problem {
	return problem{"urn:problem:validation", "Validation failed", http.StatusBadRequest, "The request is invalid."}
}

func notFoundProblem() problem {
	return problem{"urn:problem:not-found", "Task not found", http.StatusNotFound, "The requested task does not exist."}
}

func internalProblem() problem {
	return problem{"urn:problem:internal", "Internal server error", http.StatusInternalServerError, "The server could not complete the request."}
}

func preconditionRequiredProblem() problem {
	return problem{"urn:problem:precondition-required", "Precondition required", http.StatusPreconditionRequired, "A valid If-Match header is required."}
}

func preconditionFailedProblem() problem {
	return problem{"urn:problem:precondition-failed", "Precondition failed", http.StatusPreconditionFailed, "The supplied task version is stale."}
}

func writeTask(w http.ResponseWriter, status int, value task.Task) {
	w.Header().Set("ETag", strconv.Quote(strconv.FormatInt(value.Version, 10)))
	writeJSON(w, status, responseFromTask(value))
}

func writeProblem(w http.ResponseWriter, value problem) {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(value.Status)
	_ = json.NewEncoder(w).Encode(value)
}

func writeJSON(w http.ResponseWriter, status int, value any) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)
	_ = json.NewEncoder(w).Encode(value)
}